requirements.txt
asgiref==3.2.5
certifi==2019.11.28
Django==3.0.3
django-automated-logging==5.0.1
django-taggit==1.2.0
pytz==2019.3
sqlparse==0.3.0
